+++ src/main/java/com/jsl/oa/utils/Processing.java
@@ -434,7 +434,7 @@ public class Processing {
      * @Date: 2024/4/11
      * @Param category:
      **/
-    public static String turnReviewCategory(Integer category) {
+    public static String turnReviewCategory(short category) {
         switch (category) {
             case 0:
               return "子系统";
@@ -445,7 +445,7 @@ public class Processing {
         }
     }
 
-    public static String turnReviewResult(Integer result) {
+    public static String turnReviewResult(short result) {
         switch (result) {
             case 0:
                 return "已拒绝";
